Separate hourly charts into their own component for compare page
