+++ src/components/index.js
@@ -32,3 +32,4 @@ export TextWithBackground from './TextWithBackground/TextWithBackground';
 export SearchSelect from './SearchSelect/SearchSelect';
 export CompareInputPanel from './CompareInputPanel/CompareInputPanel';
 export CompareTimeSeriesCharts from './CompareTimeSeriesCharts/CompareTimeSeriesCharts';
+export CompareHourCharts from './CompareHourCharts/CompareHourCharts';
